Phase F-1 wave 3: extract Idea/Variant/Event/Experiment mixins (#114)
- `_ops/ideas.py` — `_IdeaOpsMixin` (read_idea, list_ideas,
  create_idea, mark_idea_ready; 45 SLOC).
- `_ops/variants.py` — `_VariantOpsMixin` (read_variant, list_variants,
  create_variant, declare_variant_evaluation_error, integrate_variant;
  78 SLOC).
- `_ops/events.py` — `_EventOpsMixin` (events, replay, read_range) plus
  the module-level `iter_events_by_type`; `__init__.py` now re-exports
  it from `._ops.events` (21 SLOC).
- `_ops/experiment.py` — `_ExperimentOpsMixin` (experiment-state,
  dispatch-mode, policy-error, checkpoint export/import, and the
  evaluation-schema validator `validate_evaluation` /
  `_validate_evaluation`; 176 SLOC). The deferred checkpoint imports
  are rewritten `._checkpoint` -> `.._checkpoint` (now one level up
  from `_ops/`, per plan §8.6); export/import_checkpoint `cast` `self`
  to `_StoreBase` because `_checkpoint` needs the full composite
  surface (it calls both `_StoreCore` primitives and `_WorkerOpsMixin`
  credential helpers) — plan §8.2 option (b).
- `_StoreCore` gains a `_validate_evaluation` abstract stub for the
  wave-4 `_TaskOpsMixin` callers; MRO routes to `_ExperimentOpsMixin`.
- Composite widened to 6 mixins + `_StoreCore`; task ops stay inline.

Behavior-preserving: storage suite 373 passed / 185 skipped, full
pyright clean, complexity gate green (annotation still present).
